% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-class.R
\name{alpha}
\alias{alpha}
\title{Build alpha-class objects from their components.}
\usage{
alpha(...)
}
\arguments{
\item{...}{One or more components}
}
\value{
An object of class alpha.
}
\description{
\code{alpha()} is a constructor method
}
\seealso{
\code{\link{breakaway}}
}
