% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-class.R
\name{alpha_estimate}
\alias{alpha_estimate}
\title{alpha_estimate}
\usage{
alpha_estimate(estimate = NULL, error = NULL, estimand = NULL,
  name = NULL, interval = NULL, interval_type = NULL, type = NULL,
  model = NULL, warnings = NULL, frequentist = NULL, parametric = NULL,
  plot = NULL, reasonable = NULL, other = NULL, ...)
}
\arguments{
\item{estimate}{The estimate}

\item{error}{The standard error in the estimate}

\item{estimand}{What is the estimate trying to estimate? (richness, Shannon...)}

\item{name}{The name of the method}

\item{interval}{An interval estimate}

\item{interval_type}{Type of interval estimate}

\item{type}{????}

\item{model}{What model is fit}

\item{warnings}{Any warnings?}

\item{frequentist}{Logical. Frequentist or Bayesian?}

\item{parametric}{Logical. Parametric or not?}

\item{plot}{A ggplot associated with the estimate}

\item{reasonable}{Is the estimate likely to be reasonable?}

\item{other}{Any other relevant objects}

\item{...}{Any other objects}
}
\value{
An object of class alpha_estimate
}
\description{
Build objects of class alpha_estimate from their components. \code{alpha_estimate()} is a constructor method
}
