% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_alpha_estimates.R
\name{alpha_estimates}
\alias{alpha_estimates}
\title{alpha_estimates}
\usage{
alpha_estimates(...)
}
\arguments{
\item{...}{Objects of class alpha_estimate, or a list of objects of class alpha_estimate}
}
\value{
An object of class alpha_estimates
}
\description{
Build objects of class alpha_estimates from their components. \code{alpha_estimates()} is a constructor method
}
