\name{apples}
\alias{apples}
\docType{data}
\title{``apples'' dataset}
\description{Frequency count table for a microbial community sampled in the soil of an apple orchard in Switzerland.}
\usage{apples}
\source{Walsh et. al. (2014). Restricted streptomycin use in apple orchards did not adversely alter the soil bacteria communities. \emph{Frontiers in Microbiology}, \bold{4}.}
\seealso{ \code{\link{breakaway}} }
\examples{breakaway(apples)}
\keyword{datasets}
