% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{atacama}
\alias{atacama}
\title{A canonical QIIME2 dataset}
\format{
character vector
}
\usage{
atacama
}
\description{
Data used for internal testing
retrieved from https://raw.githubusercontent.com/paulinetrinh/data/master/otu_table_atacama.txt
}
\keyword{datasets}
