% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_chao1_bc.R
\name{chao1_bc}
\alias{chao1_bc}
\title{Bias-corrected Chao1 species richness estimator}
\usage{
chao1_bc(input_data, output = NULL, answers = NULL)
}
\arguments{
\item{input_data}{An input type that can be processed by \code{convert()} or a \code{phyloseq} object}

\item{output}{Deprecated; only for backwards compatibility}

\item{answers}{Deprecated; only for backwards compatibility}
}
\value{
An object of class \code{alpha_estimate}, or \code{alpha_estimates} for \code{phyloseq} objects
}
\description{
This function implements the bias-corrected Chao1 richness estimate.
}
\note{
The authors of this package strongly discourage the use of this
estimator.  It is only valid when you wish to assume that every taxa has
equal probability of being observed. You don't really think that's possible,
do you?
}
\examples{


chao1_bc(apples)


}
\author{
Amy Willis
}
