% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_chao_shen.R
\name{chao_shen}
\alias{chao_shen}
\title{The Chao-Shen estimate of Shannon diversity}
\usage{
chao_shen(input_data)
}
\arguments{
\item{input_data}{An input type that can be processed by \code{convert()}}
}
\value{
An object of class \code{alpha_estimate}
}
\description{
The Chao-Shen estimate of Shannon diversity
}
