% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_convert.R
\name{check_format}
\alias{check_format}
\title{Run some basic checks on a possible frequency count table}
\usage{
check_format(output_data)
}
\value{
A checked frequency count table
}
\description{
Run some basic checks on a possible frequency count table
}
