% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_wrap.R
\name{cutoff_wrap}
\alias{cutoff_wrap}
\title{Wrapper for phyloseq objects}
\usage{
cutoff_wrap(data)
}
\arguments{
\item{data}{Frequency count table}
}
\value{
Cutoff value
}
\description{
Wrapper for phyloseq objects
}
