% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{dv}
\alias{dv}
\title{DivNet model fitted to soil_phylo dataset}
\format{
An object of class `diversityEstimates`

An object of class `diversityEstimates`
}
\usage{
dv

dv
}
\description{
DivNet model fit via call
#dv <- DivNet::divnet(soil_phylum, X = NULL)
Further details in diversity hypothesis testing vignette

DivNet model fit via call
#dv <- DivNet::divnet(soil_phylum, X = NULL)
}
\references{
Whitman, T., Pepe-Ranney, C., Enders, A., Koechli, C., Campbell, A.,  Buckley, D. H., Lehmann, J. (2016). \emph{Dynamics of microbial community composi-tion and soil organic carbon mineralization in soil following addition of pyrogenic andfresh organic matter}. The ISME journal, 10(12):2918. <doi: 10.1038/ismej.2016.68>.

Whitman, T., Pepe-Ranney, C., Enders, A., Koechli, C., Campbell, A.,  Buckley, D. H., Lehmann, J. (2016). \emph{Dynamics of microbial community composi-tion and soil organic carbon mineralization in soil following addition of pyrogenic andfresh organic matter}. The ISME journal, 10(12):2918. <doi: 10.1038/ismej.2016.68>.
}
\keyword{datasets}
