% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{dv_days_1_2}
\alias{dv_days_1_2}
\title{DivNet model fitted to \code{soil_phylo} dataset including observations on day 1}
\format{
An object of class \code{diversityEstimates}
}
\usage{
dv_days_1_2
}
\description{
DivNet model fit via call
#dv_days_1_2 <- DivNet::divnet(soil_phylum_days_1_2, X = NULL)
Further details in diversity hypothesis testing vignette
}
\references{
Whitman, T., Pepe-Ranney, C., Enders, A., Koechli, C., Campbell, A.,  Buckley, D. H., Lehmann, J. (2016). \emph{Dynamics of microbial community composition and soil organic carbon mineralization in soil following addition of pyrogenic and fresh organic matter}. The ISME journal, 10(12):2918. <doi: 10.1038/ismej.2016.68>.
}
\keyword{datasets}
