% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_core.R
\name{frequency_count_or_proportion_or_column}
\alias{frequency_count_or_proportion_or_column}
\title{What type of data is this?}
\usage{
frequency_count_or_proportion_or_column(input)
}
\arguments{
\item{input}{The input data}
}
\value{
A string telling you what you had
}
\description{
What type of data is this?
}
