% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F_stat.R
\name{get_F_stat}
\alias{get_F_stat}
\title{Calculate F statistic under null hypothesis LB = 0 using output from betta()
or betta_random()}
\usage{
get_F_stat(fitted_betta, L)
}
\arguments{
\item{fitted_betta}{A fitted betta object -- i.e., the output of either betta() or
betta_random() -- containing fixed effect estimates of interest.}

\item{L}{An m x p matrix defining the null LB = 0. L must have full row rank.}
}
\value{
The calculated F-statistic
}
\description{
This function calculates an F statistic for a test of null hypothesis LB = 0
(against an unrestricted alternative) where B is a
vector of p fixed effects returned by betta() or betta_random() and L is an
m x p matrix with linearly independent rows.
}
