% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_good_turing.R
\name{good_turing}
\alias{good_turing}
\title{The Good-Turing estimate of Shannon diversity}
\usage{
good_turing(input_data)
}
\arguments{
\item{input_data}{An input type that can be processed by \code{convert()} or a \code{phyloseq} object}
}
\value{
An object of class \code{alpha_estimate}, or \code{alpha_estimates} for \code{phyloseq} objects
}
\description{
The Good-Turing estimate of Shannon diversity
}
