\name{hawaii}
\alias{hawaii}
\docType{data}
\title{``hawaii'' dataset}
\description{Frequency count table for a microbial community sampled in the soil near the Kohala Volcano, Hawaii.}
\usage{hawaii}
\source{Chadwick, O.A. et al. (2007). Precontact vegetation and soil nutrient status in the shadow of Kohala Volcano, Hawaii, \emph{Geomorphology}, \bold{89} 1, 70-83.}
\seealso{ \code{\link{breakaway}} }
\examples{breakaway(hawaii)}
\keyword{datasets}
