% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_core.R
\name{hill}
\alias{hill}
\title{Plug-in Hill numbers}
\usage{
hill(input, q)
}
\arguments{
\item{input}{A frequency count table, vector of abundances, or vector of
proportions.}

\item{q}{The Hill number of interest. q = 0 corresponds to species richness, q = 2 corresponds to inverse Simpson, etc.}
}
\value{
The Hill number of the population given by input.
}
\description{
Plug-in Hill numbers, for population-level data.
}
\note{
This function is intended for population-level data. If you are
dealing with a microbial sample, use DivNet instead.
}
