% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_core.R
\name{hill_se}
\alias{hill_se}
\title{alpha diversity std errors}
\usage{
hill_se(Cest, Cse, q, props, Dest, c)
}
\arguments{
\item{Cest}{TODO}

\item{Cse}{TODO}

\item{q}{TODO}

\item{props}{TODO}

\item{Dest}{TODO}

\item{c}{TODO}
}
\description{
TODO
}
