% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_covariate_matrix.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Make design matrix}
\usage{
make_design_matrix(phyloseq_object, variables)
}
\arguments{
\item{phyloseq_object}{A phyloseq object}

\item{variables}{variable names}
}
\value{
A matrix object giving the design matrix from the desired variables.
}
\description{
Make design matrix
}
