% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_objective_bayes.R
\name{objective_bayes_geometric}
\alias{objective_bayes_geometric}
\title{Estimate species richness with an objective Bayes method using a geometric model}
\usage{
objective_bayes_geometric(
  data,
  output = TRUE,
  plot = TRUE,
  answers = FALSE,
  tau = 10,
  burn.in = 100,
  iterations = 2500,
  Metropolis.stdev.N = 75,
  Metropolis.start.theta = 1,
  Metropolis.stdev.theta = 0.3
)
}
\arguments{
\item{data}{A \code{\link[base]{data.frame}} or tibble containing the frequency of frequencies data.}

\item{output}{A logical value indicating whether the function should return the full posterior samples and diagnostic plots. Defaults to \code{TRUE}.}

\item{plot}{A logical value indicating whether to print a plot of the Markov Chain Monte Carlo diagnostic chains and posterior distribution. Defaults to \code{TRUE}.}

\item{answers}{A logical value indicating whether to return print-ready summary statistics and credible intervals. Defaults to \code{FALSE}.}

\item{tau}{An integer indicating the maximum frequency of the data to be analyzed. Defaults to \code{10}.}

\item{burn.in}{An integer indicating the number of burn-in iterations for the Markov Chain Monte Carlo sampler. Defaults to \code{100}.}

\item{iterations}{An integer indicating the number of additional iterations for the Markov Chain Monte Carlo sampler after the burn-in iterations. Defaults to \code{2500}.}

\item{Metropolis.stdev.N}{A numeric value specifying the standard deviation for the Metropolis-Hastings proposal distribution of the parameter N. Defaults to \code{75}.}

\item{Metropolis.start.theta}{Starting value in the Metropolis algorithm for the parameter theta. Defaults to \code{1}.}

\item{Metropolis.stdev.theta}{A numeric value specifying the standard deviation for the Metropolis-Hastings proposal distribution of the parameter theta. Defaults to \code{0.3}.}
}
\value{
A list of results, including \item{est}{the median of estimates of N}, \item{ci}{a confidence interval for N},
\item{mean}{the mean of estimates of N}, \item{semeanest}{the standard error of mean estimates},
\item{dic}{the DIC of the model}, \item{fits}{fitted values}, and \item{diagnostics}{model diagonstics}.
}
\description{
Estimate species richness with an objective Bayes method using a geometric model
}
