% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{pasolli_et_al}
\alias{pasolli_et_al}
\title{(Data) Data frame of covariate information about pasolli_et_al.}
\format{
A data frame with 4930 rows and 9 variables:
\describe{
  \item{SGB.ID}{sample ID}
  \item{AverageAbundance}{average abundance in the sample}
  \item{std}{standard deviation}
  \item{q1}{first quartile}
  \item{MedianAbundance}{median abundance}
  \item{q3}{third quartile}
  \item{min}{minimum}
  \item{max}{maximum}
  \item{#.Samples}{number of samples}
  ...
}
}
\usage{
pasolli_et_al
}
\description{
(Data) Data frame of covariate information about pasolli_et_al.
}
\keyword{datasets}
