% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_wrap.R
\name{physeq_wrap}
\alias{physeq_wrap}
\title{Wrapper for phyloseq objects}
\usage{
physeq_wrap(fn, physeq, ...)
}
\arguments{
\item{fn}{alpha diversity estimator function with \code{breakaway} to be applied to \code{physeq}}

\item{physeq}{A \code{phyloseq} object, or an object of class \code{otu_table}}

\item{...}{Additional arguments for fn}
}
\value{
Object of class \code{alpha_estimates}
}
\description{
Wrapper for phyloseq objects
}
