% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_estimates.R
\name{plot.alpha_estimates}
\alias{plot.alpha_estimates}
\title{Plot function for alpha_estimates class}
\usage{
\method{plot}{alpha_estimates}(x, ...)
}
\arguments{
\item{x}{Object of class \code{alpha_estimates}}

\item{...}{See details}
}
\description{
Plot function for alpha_estimates class
}
\details{
... Can currently include "fullCI", a boolean to indicate whether you want the plot to include the full confidence intervals. Defaults to FALSE.
}
\examples{
\dontrun{
library(phyloseq)
data(GlobalPatterns)
alphas <- breakaway(GlobalPatterns)
plot(alphas)
}
}
