% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_integration.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Plot alpha diversity with less incorrect error bars}
\usage{
plot_alpha(physeq, x = "samples", color = NULL, shape = NULL, nrow = 1,
  measures = NULL, estimators = NULL, upper_limits = NULL)
}
\arguments{
\item{physeq}{(Required). \code{\link{phyloseq-class}}, or alternatively,
an \code{\link{otu_table-class}}. The data about which you want to estimate.}

\item{x}{(Optional). A variable to map to the horizontal axis. The vertical
 axis will be mapped to the alpha diversity index/estimate
 and have units of total taxa, and/or index value (dimensionless).
 This parameter (\code{x}) can be either a character string indicating a
 variable in \code{sample_data}
 (among the set returned by \code{sample_variables(physeq)} );
 or a custom supplied vector with length equal to the number of samples
 in the dataset (nsamples(physeq)).

 The default value is \code{"samples"}, which will map each sample's name
 to a separate horizontal position in the plot.}

\item{color}{(Optional). Default \code{NULL}.
The sample variable to map to different colors.
Like \code{x}, this can be a single character string of the variable name in
\code{sample_data}
(among the set returned by \code{sample_variables(physeq)} );
or a custom supplied vector with length equal to the number of samples
in the dataset (nsamples(physeq)).
The color scheme is chosen automatically by \code{link{ggplot}},
but it can be modified afterward with an additional layer using
\code{\link[ggplot2]{scale_color_manual}}.}

\item{shape}{(Optional). Default \code{NULL}. The sample variable to map
to different shapes. Like \code{x} and \code{color},
this can be a single character string
of the variable name in
\code{sample_data}
(among the set returned by \code{sample_variables(physeq)} );
or a custom supplied vector with length equal to the number of samples
in the dataset (nsamples(physeq)).
The shape scale is chosen automatically by \code{link{ggplot}},
but it can be modified afterward with an additional layer using
\code{\link[ggplot2]{scale_shape_manual}}.}

\item{nrow}{(Optional). Default is \code{1},
meaning that all plot panels will be placed in a single row,
side-by-side.
This argument is passed to \code{\link[ggplot2]{facet_wrap}}.
If \code{NULL}, the number of rows and columns will be
chosen automatically (wrapped) based on the number of panels
and the size of the graphics device.}

\item{measures}{(Optional). Default is \code{NULL}, meaning that
all available alpha-diversity measures will be included in plot panels.
Alternatively, you can specify one or more measures
as a character vector of measure names.
Values must be among those supported:
\code{c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher")}.}

\item{title}{(Optional). Default \code{NULL}. Character string.
The main title for the graphic.}
}
\value{
A \code{\link{ggplot}} plot object summarizing
 the richness estimates, and their standard error.
}
\description{
For no good reason, we almost never plot alpha diversity with its appropriate
standard errors. This function is intended to fix that.
}
\details{
NOTE: This code is modify from the phyloseq functions plot_richness
 and estimate_richness. Please cite phyloseq as well as breakaway
 if you use this function!
}
\seealso{
\code{\link{estimate_richness}}
}
