% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_poisson_nof1.R
\name{poisson_model_nof1}
\alias{poisson_model_nof1}
\title{PoissonModelNof1}
\usage{
poisson_model_nof1(input_data, cutoff = 10)
}
\arguments{
\item{input_data}{A frequency count table}

\item{cutoff}{The largest frequency to use for predicting f0 and f1. Defaults to 10.}
}
\value{
An object of class \code{alpha_estimate} containing a summary of the
fitted model
}
\description{
A model to estimate the number of missing taxa under a zero- and one-truncated Poisson Model
}
