% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_build.R
\name{proportions_instead}
\alias{proportions_instead}
\title{OTU table to relative abundances}
\usage{
proportions_instead(the_table)
}
\arguments{
\item{the_table}{An OTU table}
}
\value{
A table of relative abundances
}
\description{
OTU table to relative abundances
}
