% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_resample.R
\name{resample_estimate}
\alias{resample_estimate}
\title{Resampling-based estimates of alpha diversity}
\usage{
resample_estimate(my_data, my_function, my_sample_size = NA, ...)
}
\arguments{
\item{my_data}{The columns of your otu table, giving the number of counts
for each taxa}

\item{my_function}{The alpha diversity metric that you're interested in
(e.g. Shannon)}

\item{my_sample_size}{The number of reads that you want to resample. Common
estimates of alpha diversity (such as the plug-in estimate of Shannon
diversity) are very sensitive to sample size. To gauge the effect of
differing sample sizes on the variance in estimating alpha diversity, it is
(currently) advisable to include multiple different sample sizes in your
resampling}

\item{...}{Other arguments to pass to \code{my_function}}
}
\value{
A (independent-bootstrap resampled) estimate of the alpha diversity
metric given by my_function
}
\description{
This function resamples OTUs to a specified read count, then calculates a
given alpha diversity metric based on the resample. Useful for determining
variance of alpha diversity estimates.
}
\examples{










# Bootstrap observations to calculate a re-sample estimate of the Shannon index
resample_estimate(toy_otu_table[,1], shannon)

# To see the sampling distribution of the resample estimates, repeat this 200 times
shannon_estimates <- replicate(200, resample_estimate(toy_otu_table[,1], shannon))
hist(shannon_estimates)

# To investigate the effect of sample size on the resample estimate of the Shannon index,
# re-sample over multiple sample sizes; in this case, the number of reads
ns <- apply(toy_otu_table, 2, sum)
shannon_estimates <- replicate(200, resample_estimate(toy_otu_table[,1], 
shannon, 
my_sample_size = ns))

hist(shannon_estimates)
sd(shannon_estimates) # a not too terrible estimate of the standard error in estimating Shannon










}
\author{
Amy Willis
}
