% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{rztnbinomtable}
\alias{rztnbinomtable}
\title{beta version: Zero-truncated negative binomially distributed frequency count
tables.}
\usage{
rztnbinomtable(C, size, probability)
}
\arguments{
\item{C}{species richness}

\item{size}{size parameter for the negative binomial distribution}

\item{probability}{probability parameter for the negative binomial
distribution}
}
\description{
Simulate a frequency count table based on a negative binomial model.
Zero-truncated, obviously.
}
\author{
Amy Willis
}
