% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_sample.R
\name{sample_inverse_simpson}
\alias{sample_inverse_simpson}
\title{Plug-in Inverse Simpson diversity}
\usage{
sample_inverse_simpson(input_data)
}
\arguments{
\item{input_data}{An input type that can be processed by \code{convert()} or a \code{phyloseq} object}
}
\value{
An object of class \code{alpha_estimate}, or \code{alpha_estimates} for \code{phyloseq} objects
}
\description{
This function implements the plug-in Inverse Simpson diversity
}
\examples{
sample_inverse_simpson(apples)

}
