% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_sample.R
\name{sample_richness}
\alias{sample_richness}
\title{Sample richness estimator}
\usage{
sample_richness(input_data)
}
\arguments{
\item{input_data}{An input type that can be processed by \code{convert()} or a \code{phyloseq} object}
}
\value{
An object of class \code{alpha_estimate}, or \code{alpha_estimates} for \code{phyloseq} objects
}
\description{
This function implements the sample richness estimate, which is the number of non-zero taxa per sample.
}
\examples{


sample_richness(apples)


}
