% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_tools.R
\name{sample_size_figure}
\alias{sample_size_figure}
\title{Plot the power obtained with sample size}
\usage{
sample_size_figure(control_group_est, se_est, diff = 5, samples = 20)
}
\arguments{
\item{control_group_est}{An estimate of the alpha diversity parameter for the control group}

\item{se_est}{An estimate of the (common) standard deviation}

\item{diff}{An estimate of the difference between the control and non-control groups}

\item{samples}{Number of bootstrap resamples used to estimate the sample size. Increase for a more accurate estimate.}
}
\value{
A plot of the power with the sample size
}
\description{
Plot the power obtained with sample size
}

