% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_core.R
\name{shannon}
\alias{shannon}
\title{Plug-in Shannon index}
\usage{
shannon(input)
}
\arguments{
\item{input}{A frequency count table, vector of abundances, or vector of
proportions.}
}
\value{
The Shannon index of the population given by input.
}
\description{
Plug-in Shannon index, for population-level data.
}
\note{
This function is intended for population-level data. If you are
dealing with a microbial sample, use \code{\link{shannon_better}} instead.
}
