% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_better.R
\name{shannon_better}
\alias{shannon_better}
\title{Shannon diversity estimates}
\usage{
shannon_better(input, ccc = NA, ccc_se = NA, quantile = 0.95)
}
\arguments{
\item{input}{A frequency count table or vector of abundances.}

\item{ccc}{An estimate of the total species richness of the community under
study. If none is given, \code{\link{breakaway}} is used}

\item{ccc_se}{The standard error in the estimate of ccc.}
}
\value{
An estimate, standard error, and confidence set of the index.
}
\description{
In many ecological surveys, only a small fraction of the total ecosystem is
observed. In order to draw conclusions about the total ecosystem, it is
necessary to adjust for inexhaustive sampling in estimates of alpha
diversity. This function implements such an adjustment for the Shannon
index.
}
\examples{




shannon_better(apples)




}
