% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_betta.R
\name{simulate_betta}
\alias{simulate_betta}
\title{Simulate from a fitted betta model}
\usage{
simulate_betta(fitted_betta, nsim)
}
\arguments{
\item{fitted_betta}{A fitted betta object}

\item{nsim}{Number of times to simulate}
}
\value{
A list of length nsim, each element of which is a vector of
simulated Y-values under the fitted betta model
}
\description{
Simulate from a fitted betta model
}
