% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_betta_random.R
\name{simulate_betta_random}
\alias{simulate_betta_random}
\title{Simulate from a fitted betta_random model}
\usage{
simulate_betta_random(fitted_betta, nsim)
}
\arguments{
\item{fitted_betta}{A fitted betta_random object}

\item{nsim}{Number of times to simulate}
}
\value{
A list of length \code{nsim}, each element of which is a vector of
simulated Y-values under the fitted betta model
}
\description{
Simulate from a fitted betta_random model
}
