% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{submodel_test}
\alias{submodel_test}
\title{Output from test_submodel() for use in hypothesis testing vignette}
\format{
A list containing the bootstrap p-value, the observed F-statistic,
and a vector of bootstrapped F-statistics
}
\usage{
submodel_test
}
\description{
Output from test_submodel() for use in hypothesis testing vignette
}
\keyword{datasets}
