% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_resample.R
\name{subsample_otu}
\alias{subsample_otu}
\title{Resampling the taxa of your OTU table}
\usage{
subsample_otu(my_data, my_sample_size = NA, ...)
}
\arguments{
\item{my_data}{The columns of your OTU table, giving the number of counts
for each taxa}

\item{my_sample_size}{The number of reads of the bootstrap.}

\item{...}{Other arguments, which are ignored.}
}
\value{
A independent-bootstrap resampled OTU frequencies, in frequency
count table form.
}
\description{
This function resamples OTUs to a specified read count
}
