% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_core.R
\name{to_proportions}
\alias{to_proportions}
\title{convert to proportions}
\usage{
to_proportions(input, type)
}
\arguments{
\item{input}{the input data}

\item{type}{What time}
}
\description{
TODO write as generic method
}
