% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{toy_metadata}
\alias{toy_metadata}
\title{(Data) Data frame of covariate information about toy_otu_table.}
\format{A data frame with 143 rows and 4 variables:
\describe{
  \item{Years}{Year of sampling}
  \item{bloom2}{Did the sample correspond to a bloom event?}
  \item{Period}{What season was sampled?}
  \item{Site}{Where was the sample taken from?}
  ...
}}
\usage{
toy_metadata
}
\description{
(Data) Data frame of covariate information about toy_otu_table.
}
\keyword{datasets}
