% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{toy_otu_table}
\alias{toy_otu_table}
\title{(Data) A toy OTU table. Covariate info available in `toy_metadata`.}
\format{A data frame with 448 rows and 143 columns:
\describe{
  \item{Rows:}{The abundance of each taxon...}
  \item{Columns:}{...in each sample}
  ...
}}
\usage{
toy_otu_table
}
\description{
(Data) A toy OTU table. Covariate info available in `toy_metadata`.
}
\keyword{datasets}
