% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{toy_taxonomy}
\alias{toy_taxonomy}
\title{(Data) The taxonomy of the OTUs in `toy_otu_table`.}
\format{
An object of class \code{factor} of length 448.
}
\usage{
toy_taxonomy
}
\description{
(Data) The taxonomy of the OTUs in `toy_otu_table`.
}
\keyword{datasets}
