% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_true.R
\name{true_gini}
\alias{true_gini}
\title{Calculate the true Gini-Simpson index}
\usage{
true_gini(input)
}
\arguments{
\item{input}{A vector of proportions.}
}
\value{
The Gini-Simpson index of the population given by input.
}
\description{
Calculate the true Gini-Simpson index
}
\note{
This function is intended for population-level data. If you are
dealing with a microbial sample, use DivNet instead.
}
