% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_true.R
\name{true_hill}
\alias{true_hill}
\title{Calculate the true Hill numbers}
\usage{
true_hill(input, q)
}
\arguments{
\item{input}{A vector of proportions.}

\item{q}{The Hill number of interest. q = 0 corresponds to species richness, q = 2 corresponds to inverse Simpson, etc.}
}
\value{
The Hill number of the population given by input.
}
\description{
Calculate the true Hill numbers
}
