% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_true.R
\name{true_inverse_simpson}
\alias{true_inverse_simpson}
\title{Calculate the true Inverse Simpson index}
\usage{
true_inverse_simpson(input)
}
\arguments{
\item{input}{A vector of proportions.}
}
\value{
The inverse-Simpson index of the population given by input.
}
\description{
Calculate the true Inverse Simpson index
}
\note{
This function is intended for population-level data. If you are
dealing with a microbial sample, use DivNet instead.
}
