% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_true.R
\name{true_shannon_e}
\alias{true_shannon_e}
\title{Calculate the true Shannon's equitability index}
\usage{
true_shannon_e(input)
}
\arguments{
\item{input}{A vector of proportions.}
}
\value{
The Shannon E's of the population given by input.
}
\description{
Calculate the true Shannon's equitability index
}
\note{
This function is intended for population-level data. If you are
dealing with a microbial sample, use DivNet instead.
}
