% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_true.R
\name{true_simpson}
\alias{true_simpson}
\title{Calculate the true Simpson index}
\usage{
true_simpson(input)
}
\arguments{
\item{input}{A vector of proportions.}
}
\value{
The Simpson index of the population given by input.
}
\description{
Calculate the true Simpson index
}
\note{
This function is intended for population-level data. If you are
dealing with a microbial sample, use DivNet instead.
}
