% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-class.R
\name{alpha}
\alias{alpha}
\title{alpha}
\usage{
alpha(estimate = 1000, error = 100)
}
\arguments{
\item{estimate}{the estimate}

\item{error}{the error}

\item{interval}{an interval estimate}

\item{type}{TODO}

\item{model}{TODO}

\item{warnings}{TODO}

\item{frequentist}{frequentist or Bayesian}

\item{parametric}{TODO}

\item{other}{TODO}

\item{estimate}{the estimate}
}
\value{
An object of class alpha.
}
\description{
Build alpha-class objects from their components. \code{alpha()} is a constructor method
}
