% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{alpha_better}
\alias{alpha_better}
\title{alpha diversity estimates}
\usage{
alpha_better(input, q, ccc = NA)
}
\arguments{
\item{input}{TODO}

\item{q}{TODO}

\item{ccc}{TODO}
}
\description{
TODO
}
