% Please edit documentation in R/bData.R
\docType{data}
\name{apples}
\alias{apples}
\title{(Data) Frequency count table of soil microbes in an apples orchard.}
\format{A data frame with 88 rows and 2 variables:
\describe{
  \item{index}{an index variable}
  \item{frequency}{number of taxa that were observed with this frequency}
  ...
}}
\source{
\url{http://onlinelibrary.wiley.com/doi/10.1111/biom.12332/abstract}
}
\usage{
apples
}
\description{
(Data) Frequency count table of soil microbes in an apples orchard.
}
\references{
Willis, A. and Bunge, J. (2015). Estimating diversity via
frequency ratios. \emph{Biometrics}, \bold{71}(4), 1042--1049.
\url{http://onlinelibrary.wiley.com/doi/10.1111/biom.12332/abstract}

Walsh, F. et al. (2014). (2014). Restricted streptomycin use in apple 
orchards did not adversely alter the soil bacteria communities. 
\emph{Frontiers in Microbiology} \bold{4}, 383.
}
\keyword{datasets}
