% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_chao1.R
\name{chao1}
\alias{chao1}
\title{Chao1 species richness estimator}
\usage{
chao1(input_data, output = NULL, answers = NULL)
}
\arguments{
\item{input_data}{The sample frequency count table for the population of interest.
See dataset apples for sample formatting.}

\item{output}{Deprecated; only for backwards compatibility}

\item{answers}{Deprecated; only for backwards compatibility}
}
\value{
The results of the estimator, including standard error.
}
\description{
This function implements the Chao1 richness estimate, which is often
mistakenly referred to as an index.
}
\note{
The authors of this package strongly discourage the use of this
estimator.  It is only valid when you wish to assume that every taxa has
equal probability of being observed. You don't really think that's possible,
do you?
}
\examples{


chao1(apples)


}
\author{
Amy Willis
}
