% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chao_functions.R
\name{chao1}
\alias{chao1}
\title{Chao1 species richness estimator}
\usage{
chao1(data, print = TRUE, answers = FALSE)
}
\arguments{
\item{data}{The sample frequency count table for the population of interest.
See dataset apples for sample formatting.}

\item{print}{Logical: whether the results should be printed to screen.}

\item{answers}{Should the answers be returned as a list?}
}
\value{
The results of the estimator, including standard error.
}
\description{
This function implements the Chao1 richness estimate, which is often
mistakenly referred to as an index.
}
\note{
The authors of this package strongly discourage the use of this
estimator.  It is only valid when you wish to assume that every taxa has
equal probability of being observed. You don't really think that's possible,
do you?
}
\examples{








chao1(apples)








}
\author{
Amy Willis
}
