% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_chao.R
\name{chao1_bc}
\alias{chao1_bc}
\title{Bias-corrected Chao1 species richness estimator}
\usage{
chao1_bc(data, print = TRUE, answers = FALSE)
}
\arguments{
\item{data}{The sample frequency count table for the population of interest.
See dataset apples for sample formatting.}

\item{print}{Logical: whether the results should be printed to screen.}

\item{answers}{Should the answers be returned as a list?}
}
\value{
The results of the estimator, including standard error.
}
\description{
This function implements the bias-corrected Chao1 richness estimate.
}
\note{
The authors of this package strongly discourage the use of this
estimator. It is underpinned by totally implausible assumptions that are not
made by other richness estimators.  Bias correcting Chao1 is the least of
your problems.
}
\examples{





















chao1_bc(apples)





















}
\author{
Amy Willis
}
