% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_chao_bunge.R
\name{chao_bunge}
\alias{chao_bunge}
\title{Chao-Bunge species richness estimator}
\usage{
chao_bunge(input_data, cutoff = 10, output = NULL, answers = NULL)
}
\arguments{
\item{input_data}{An input type that can be processed by \code{convert()}}

\item{cutoff}{The maximum frequency to use in fitting.}

\item{output}{Deprecated; only for backwards compatibility}

\item{answers}{Deprecated; only for backwards compatibility}
}
\value{
An object of class \code{alpha_estimate}
}
\description{
This function implements the species richness estimation procedure outlined
in Chao & Bunge (2002).
}
\examples{

chao_bunge(apples)

}
\author{
Amy Willis
}
