% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_convert.R
\name{convert}
\alias{convert}
\title{convert between different inputs for alpha-diversity estimates}
\usage{
convert(input_data)
}
\arguments{
\item{Supported}{types include filenames, data frames, matrices, vectors...}
}
\value{
Frequency count able 

Inputs slated for development include phyloseq and otu_table
}
\description{
convert between different inputs for alpha-diversity estimates
}
