% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_wrap.R
\name{cutoff_wrap}
\alias{cutoff_wrap}
\title{Find a cut-off for estimates relying on contiguous counts}
\usage{
cutoff_wrap(my_data, requested = NA)
}
\arguments{
\item{my_data}{Frequency count table}

\item{requested}{The user-requested cutoff}
}
\value{
Cutoff value
}
\description{
Find a cut-off for estimates relying on contiguous counts
}
