% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{hawaii}
\alias{hawaii}
\title{(Data) Frequency count table of soil microbes in Hawaii.}
\format{
A data frame with 198 rows and 2 variables:
\describe{
  \item{index}{an index variable}
  \item{frequency}{number of taxa that were observed with this frequency}
  ...
}
}
\usage{
hawaii
}
\description{
(Data) Frequency count table of soil microbes in Hawaii.
}
\references{
Willis, A. and Bunge, J. (2015). Estimating diversity via
frequency ratios. \emph{Biometrics}, \bold{71}(4), 1042--1049.
\url{https://doi.org/10.1111/biom.12332}
}
\keyword{datasets}
