% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_build.R
\name{make_frequency_count_table}
\alias{make_frequency_count_table}
\title{Draw frequency count subtables from an OTU table}
\usage{
make_frequency_count_table(labels)
}
\description{
Draw frequency count subtables from an OTU table
}
