% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_objective_bayes.R
\name{objective_bayes_geometric}
\alias{objective_bayes_geometric}
\title{Estimate species richness with an objective Bayes method using a geometric model}
\usage{
objective_bayes_geometric(data, output = TRUE, plot = TRUE,
  answers = FALSE, tau = 10, burn.in = 100, iterations = 2500,
  Metropolis.stdev.N = 75, Metropolis.start.theta = 1,
  Metropolis.stdev.theta = 0.3)
}
\arguments{
\item{data}{TODO}

\item{output}{TODO}

\item{plot}{TODO}

\item{answers}{TODO}

\item{tau}{TODO}

\item{burn.in}{TODO}

\item{iterations}{TODO}

\item{Metropolis.stdev.N}{TODO}

\item{Metropolis.start.theta}{TODO}

\item{Metropolis.stdev.theta}{TODO}
}
\description{
Estimate species richness with an objective Bayes method using a geometric model
}
