% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_objective_bayes.R
\name{objective_bayes_mixedgeo}
\alias{objective_bayes_mixedgeo}
\title{Objective Bayes species richness estimate with the mixed-geometric model}
\usage{
objective_bayes_mixedgeo(
  data,
  output = TRUE,
  plot = TRUE,
  answers = FALSE,
  tau = 10,
  burn.in = 100,
  iterations = 2500,
  Metropolis.stdev.N = 100,
  Metropolis.start.T1 = 1,
  Metropolis.stdev.T1 = 2,
  Metropolis.start.T2 = 3,
  Metropolis.stdev.T2 = 2,
  bars = 3
)
}
\arguments{
\item{data}{TODO(Kathryn)}

\item{output}{TODO(Kathryn)}

\item{plot}{TODO(Kathryn)}

\item{answers}{TODO(Kathryn)}

\item{tau}{TODO(Kathryn)}

\item{burn.in}{TODO(Kathryn)}

\item{iterations}{TODO(Kathryn)}

\item{Metropolis.stdev.N}{TODO(Kathryn)}

\item{Metropolis.start.T1}{TODO(Kathryn)}

\item{Metropolis.stdev.T1}{TODO(Kathryn)}

\item{Metropolis.start.T2}{TODO(Kathryn)}

\item{Metropolis.stdev.T2}{TODO(Kathryn)}

\item{bars}{TODO(Kathryn)}
}
\value{
A list of results, including \item{est}{the median of estimates of N}, \item{ci}{a confidence interval for N},
\item{mean}{the mean of estimates of N}, \item{semeanest}{the standard error of mean estimates},
\item{dic}{the DIC of the model}, \item{fits}{fitted values}, and \item{diagnostics}{model diagonstics}.
}
\description{
Objective Bayes species richness estimate with the mixed-geometric model
}
