% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_objective_bayes.R
\name{objective_bayes_negbin}
\alias{objective_bayes_negbin}
\title{OB negbin}
\usage{
objective_bayes_negbin(data, output = TRUE, plot = TRUE, answers = FALSE,
  tau = 10, burn.in = 1000, iterations = 5000, Metropolis.stdev.N = 100,
  Metropolis.start.T1 = -0.8, Metropolis.stdev.T1 = 0.01,
  Metropolis.start.T2 = 0.8, Metropolis.stdev.T2 = 0.01, bars = 5)
}
\description{
OB negbin
}
