% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_objective_bayes.R
\name{objective_bayes_negbin}
\alias{objective_bayes_negbin}
\title{Objective Bayes species richness estimate with the Negative Binomial model}
\usage{
objective_bayes_negbin(data, output = TRUE, plot = TRUE,
  answers = FALSE, tau = 10, burn.in = 1000, iterations = 5000,
  Metropolis.stdev.N = 100, Metropolis.start.T1 = -0.8,
  Metropolis.stdev.T1 = 0.01, Metropolis.start.T2 = 0.8,
  Metropolis.stdev.T2 = 0.01, bars = 5)
}
\arguments{
\item{data}{TODO(Kathryn)}

\item{output}{TODO(Kathryn)}

\item{plot}{TODO(Kathryn)}

\item{answers}{TODO(Kathryn)}

\item{tau}{TODO(Kathryn)}

\item{burn.in}{TODO(Kathryn)}

\item{iterations}{TODO(Kathryn)}

\item{Metropolis.stdev.N}{TODO(Kathryn)}

\item{Metropolis.start.T1}{TODO(Kathryn)}

\item{Metropolis.stdev.T1}{TODO(Kathryn)}

\item{Metropolis.start.T2}{TODO(Kathryn)}

\item{Metropolis.stdev.T2}{TODO(Kathryn)}

\item{bars}{TODO(Kathryn)}
}
\description{
Objective Bayes species richness estimate with the Negative Binomial model
}
