% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_objective_bayes.R
\name{objective_bayes_poisson}
\alias{objective_bayes_poisson}
\title{Objective Bayes species richness estimate with the Poisson model}
\usage{
objective_bayes_poisson(
  data,
  output = TRUE,
  plot = TRUE,
  answers = FALSE,
  tau = 10,
  burn.in = 100,
  iterations = 2500,
  Metropolis.stdev.N = 75,
  Metropolis.start.lambda = 1,
  Metropolis.stdev.lambda = 0.3,
  bars = 5
)
}
\arguments{
\item{data}{TODO(Kathryn)}

\item{output}{TODO(Kathryn)}

\item{plot}{TODO(Kathryn)}

\item{answers}{TODO(Kathryn)}

\item{tau}{TODO(Kathryn)}

\item{burn.in}{TODO(Kathryn)}

\item{iterations}{TODO(Kathryn)}

\item{Metropolis.stdev.N}{TODO(Kathryn)}

\item{Metropolis.start.lambda}{TODO(Kathryn)}

\item{Metropolis.stdev.lambda}{TODO(Kathryn)}

\item{bars}{TODO(Kathryn)}
}
\description{
Objective Bayes species richness estimate with the Poisson model
}
