% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_wrap.R
\name{physeq_wrap}
\alias{physeq_wrap}
\title{Wrapper for phyloseq objects}
\usage{
physeq_wrap(input_data)
}
\arguments{
\item{input_data}{\code{phyloseq} object}
}
\value{
Object of class \code{alpha_estimates}
}
\description{
Wrapper for phyloseq objects
}
