% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_estimates.R
\name{plot.alpha_estimates}
\alias{plot.alpha_estimates}
\title{Plot function for alpha_estimates class}
\usage{
\method{plot}{alpha_estimates}(x, method = NULL, fullCI = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{alpha_estimates}.}

\item{method}{Optional. If there are multiple richness methods included in \code{x}, this can be set to the the desired method to be plotted. Defaults to the method of the first alpha diversity estimate.}

\item{fullCI}{Optional boolean. Indicates whether you want the plot to include the full confidence intervals. Defaults to \code{FALSE}.}

\item{...}{See details}
}
\description{
Plot function for alpha_estimates class
}
\details{
... does not currently have any implemented options. Optional arguments currently include "fullCI", a  Optional
}
\examples{
\dontrun{
library(phyloseq)
data(GlobalPatterns)
alphas <- breakaway(GlobalPatterns)
plot(alphas)
}
}
