% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bData.R
\docType{data}
\name{toy_otu_table}
\alias{toy_otu_table}
\title{(Data) A toy OTU table.}
\format{An object of class \code{data.frame} with 448 rows and 143 columns.}
\usage{
toy_otu_table
}
\description{
Covariate info available in `toy_metadata`. A data frame with 448 rows and 143 columns.
Rows give the abundance of each taxon; columns give the samples
}
\keyword{datasets}
